--- blocks/brand/brand.css
@@ -1,92 +1,0 @@
-.brand {
-    padding-left: var(--gutter);
-    padding-right: var(--gutter);
-    max-width: var(--max-width-block);
-    margin-left: auto;
-    margin-right: auto;
-    padding-top: var(--section-spacing-large);
-    padding-bottom: var(--section-spacing-large);
-}
-
-.brand .icon {
-    max-height: 50px;
-    transition: filter 0.3s;
-}
-
-.brand a .icon:not(.icon-stock-logo-dark):hover {
-    filter: brightness(0.7);
-}
-
-.brand .icon.icon-adobe-logo-vertical {
-    max-height: 60px;
-}
-
-.brand .icon.icon-stock-logo-dark {
-    margin-right: 70px;
-}
-
-.brand > div {
-    display: flex;
-    align-items: flex-end;
-    justify-content: center;
-    width: 100%;
-}
-
-@media screen and (max-width: 799px) {
-    .brand {
-        padding-top: var(--section-spacing);
-        padding-bottom: var(--section-spacing);
-    }
-    .brand > div {
-        flex-wrap: wrap;
-    }
-    .brand .icon.icon-stock-logo-dark {
-        margin-right: 0;
-    }
-    
-    .brand > div > div {
-        margin: 18px 0;
-    }
-    .brand > div > div:first-child {
-        margin-top: 0;
-    }
-    .brand > div > div:last-child {
-        margin-bottom: 0;
-        padding-top: 10px;
-    }
-}
-
-.brand > div > div {
-    width: 100%;
-    text-align: center;
-}
-
-.branda,
-.brandimg,
-.brandsvg {
-    display: inline-block;
-    width: auto;
-    height: auto;
-    min-width: 0;
-    margin: 0;
-    min-height: 50px;
-}
-
-.brandp.icon-container {
-    padding-top: 8px;
-}
-
-.brandp.icon-container, 
-.brandp.icon-container a {
-    display: flex;
-    justify-content: center;
-    align-items: flex-end;
-    color: #9f9f9f !important;
-    padding-bottom: 0;
-}
-
-.brandp.icon-container svg.icon {
-    width: 40px;
-    height: auto;
-    margin: 0 8px;
-}
