+++ blocks/images/images.css
@@ -1,5 +1,5 @@
 main .images {
-    max-width: var(--max-width-block);
+    max-width: var(--grid-container-width);
     margin-left: auto;
     margin-right: auto;
     display: inline-block;
